refactor: move tests from utils/fhe.rs to tests/fhe.rs
